INC-1321: Moved more KPI logic into the KPI repository
The KPIs are calculated using the 2 existing tables (reviews and
next review dates). It still makes more sense to have the queries
to get KPI-related figures inside the KPI repository even tho
these are not touching the `kpi` DB table directly.
