+++ src/test/kotlin/uk/gov/justice/digital/hmpps/incentivesapi/jpa/repository/KpiRepositoryTest.kt
@@ -1,5 +1,6 @@
 package uk.gov.justice.digital.hmpps.incentivesapi.jpa.repository
 
+import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
@@ -14,6 +15,7 @@ import org.springframework.security.test.context.support.WithMockUser
 import org.springframework.test.context.ActiveProfiles
 import uk.gov.justice.digital.hmpps.incentivesapi.helper.TestBase
 import uk.gov.justice.digital.hmpps.incentivesapi.jpa.Kpi
+import uk.gov.justice.digital.hmpps.incentivesapi.jpa.PrisonerIepLevel
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.temporal.ChronoUnit
@@ -25,14 +27,22 @@ class KpiRepositoryTest : TestBase() {
   @Autowired
   lateinit var repository: KpiRepository
 
+  @Autowired
+  lateinit var reviewsRepository: PrisonerIepLevelRepository
+
+  @Autowired
+  lateinit var nextReviewDateRepository: NextReviewDateRepository
+
   @BeforeEach
   fun setUp(): Unit = runBlocking {
     repository.deleteAll()
+    reviewsRepository.deleteAll()
   }
 
   @AfterEach
   fun tearDown(): Unit = runBlocking {
     repository.deleteAll()
+    reviewsRepository.deleteAll()
   }
 
   @Test
@@ -95,5 +105,60 @@ class KpiRepositoryTest : TestBase() {
     assertThat(updatedRecord!!.overdueReviews).isEqualTo(42)
     assertThat(updatedRecord!!.previousMonthReviewsConducted).isEqualTo(30000)
     assertThat(updatedRecord!!.previousMonthPrisonersReviewed).isEqualTo(20000)
+  }
+
+  @Test
+  fun `get number of reviews conducted and prisoners reviewed`(): Unit = runBlocking {
+    reviewsRepository.saveAll(
+      listOf(
+        // Review in matching date range
+        PrisonerIepLevel(
+          iepCode = "BAS",
+          prisonId = "LEI",
+          locationId = "LEI-1-1-001",
+          bookingId = 111111,
+          current = true,
+          reviewedBy = "TEST_STAFF1",
+          reviewTime = LocalDateTime.of(2023, 7, 1, 12, 0),
+          prisonerNumber = "A1111AA",
+        ),
+        // Review in matching date range, same prisoner
+        PrisonerIepLevel(
+          iepCode = "STD",
+          prisonId = "LEI",
+          locationId = "LEI-1-1-001",
+          bookingId = 111111,
+          current = false,
+          reviewedBy = "TEST_STAFF1",
+          reviewTime = LocalDateTime.of(2023, 7, 8, 12, 0),
+          prisonerNumber = "A1111AA",
+        ),
+        // Review in matching date range, another prisoner
+        PrisonerIepLevel(
+          iepCode = "STD",
+          prisonId = "LEI",
+          locationId = "LEI-1-1-002",
+          bookingId = 222222,
+          current = true,
+          reviewedBy = "TEST_STAFF1",
+          reviewTime = LocalDateTime.of(2023, 7, 20, 12, 0),
+          prisonerNumber = "B2222BB",
+        ),
+        // Review OUTSIDE matching date range - NOT counted
+        PrisonerIepLevel(
+          iepCode = "STD",
+          prisonId = "LEI",
+          locationId = "LEI-1-1-999",
+          bookingId = 999999,
+          current = true,
+          reviewedBy = "TEST_STAFF1",
+          reviewTime = LocalDateTime.of(2023, 6, 30, 12, 0),
+          prisonerNumber = "Z9999ZZ",
+        ),
+      ),
+    ).toList()
+
+    assertThat(repository.getNumberOfReviewsConductedAndPrisonersReviewed(LocalDate.of(2023, 8, 2)))
+      .isEqualTo(ReviewsConductedPrisonersReviewed(3, 2))
   }
 }
